# #593 - Faster vertices_list implementation for flat hyperrectangles