--- a/src/AbstractHyperrectangle.jl
+++ b/src/AbstractHyperrectangle.jl
@@ -117,7 +117,8 @@ end
 
 
 """
-    vertices_list(H::AbstractHyperrectangle{N})::Vector{Vector{N}} where {N<:Real}
+    vertices_list(H::AbstractHyperrectangle{N}
+                 )::Vector{Vector{N}} where {N<:Real}
 
 Return the list of vertices of a hyperrectangular set.
 
@@ -134,12 +135,35 @@ duplicate vertices.
 ### Notes
 
 For high dimensions, it is preferable to develop a `vertex_iterator` approach.
+
+### Algorithm
+
+First we identify the dimensions where `H` is flat, i.e., its radius is zero.
+We also compute the number of vertices that we have to create.
+
+Next we create the vertices.
+We do this by enumerating all vectors `v` of length `n` (the dimension of `H`)
+with entries `-1`/`0`/`1` and construct the corresponding vertex as follows:
+
+```math
+    \\text{vertex}(v)(i) = \\begin{cases} c(i) + r(i) & v(i) = 1 \\\\
+                                          c(i) & v(i) = 0 \\\\
+                                          c(i) - r(i) & v(i) = -1. \\end{cases}
+```
+
+For enumerating the vectors `v`, we modify the current `v` from left to right by
+changing entries `-1` to `1`, skipping entries `0`, and stopping at the first
+entry `1` (but changing it to `-1`).
+This way we only need to change the vertex in those dimensions where `v` has
+changed, which usually is a smaller number than `n`.
 """
 function vertices_list(H::AbstractHyperrectangle{N}
                       )::Vector{Vector{N}} where {N<:Real}
     n = dim(H)
 
-    # three-valued vector; initially 0 if radius is zero and 1 otherwise
+    # identify flat dimensions and store them in a binary vector whose entry in
+    # dimension i is 0 if the radius is zero and 1 otherwise
+    # the vector will later also contain entries -1
     trivector = Vector{Int8}(undef, n)
     m = 1
     c = center(H)