--- a/test/unit_Hyperrectangle.jl
+++ b/test/unit_Hyperrectangle.jl
@@ -180,11 +180,15 @@ for N in [Float64, Rational{Int}, Float32]
     @test P isa HPolytope # in 4D and for invertible map we get an HPolytope; see #631 and #1093
 
     # check that vertices_list is computed correctly if the hyperrectangle
-    # is "degenerate" in the sense that its radius is zero in all dimensions
-    # this test would take very long if all 2^100 vertices are computed (see #92)
-    H = Hyperrectangle(fill(N(1.), 100), fill(N(0.), 100))
+    # is "degenerate"/flat, i.e., its radius contains zeros
+    # these tests would crash if all 2^100 vertices were computed (see #92)
+    H = Hyperrectangle(ones(N, 100), zeros(N, 100))
     vl = vertices_list(H)
     @test vl == [H.center]
+    r = zeros(N, 100); r[1] = N(1)
+    H = Hyperrectangle(fill(N(1), 100), r)
+    vl = vertices_list(H)
+    @test ispermutation(vl, [H.center + r, H.center - r])
 
     # transform hyperrectangle into a polygon
     H1pol = convert(HPolygon, H1)